+++ packages/insomnia/src/ui/components/request-url-bar.tsx
@@ -18,8 +18,8 @@ import { convert } from '../../utils/importers/convert';
 import { invariant } from '../../utils/invariant';
 import { buildQueryStringFromParams, joinUrlAndQueryString } from '../../utils/url/querystring';
 import { SegmentEvent } from '../analytics';
-import { useCurlReadyState } from '../context/websocket-client/use-ws-ready-state';
 import { updateRequestMetaByParentId } from '../hooks/create-request';
+import { useCurlReadyState } from '../hooks/use-ready-state';
 import { useTimeoutWhen } from '../hooks/useTimeoutWhen';
 import { selectActiveEnvironment, selectActiveRequest, selectActiveWorkspace, selectHotKeyRegistry, selectResponseDownloadPath, selectSettings } from '../redux/selectors';
 import { Dropdown, DropdownButton, type DropdownHandle, DropdownItem, DropdownSection, ItemContent } from './base/dropdown';
