+++ packages/insomnia/src/ui/components/websockets/realtime-response-pane.tsx
@@ -9,7 +9,7 @@ import { ResponseTimelineEntry } from '../../../main/network/libcurl-promise';
 import { WebSocketEvent } from '../../../main/network/websocket';
 import { Response } from '../../../models/response';
 import { WebSocketResponse } from '../../../models/websocket-response';
-import { useRealtimeConnectionEvents } from '../../context/websocket-client/use-ws-connection-events';
+import { useRealtimeConnectionEvents } from '../../hooks/use-realtime-connection-events';
 import { selectActiveResponse } from '../../redux/selectors';
 import { PanelContainer, TabItem, Tabs } from '../base/tabs';
 import { ResponseHistoryDropdown } from '../dropdowns/response-history-dropdown';
@@ -116,12 +116,12 @@ const RealtimeActiveResponsePane: FC<{ requestId: string; response: WebSocketRes
 }) => {
   const [selectedEvent, setSelectedEvent] = useState<CurlEvent | WebSocketEvent | null>(null);
   const [timeline, setTimeline] = useState<ResponseTimelineEntry[]>([]);
-
   const searchInputRef = useRef<HTMLInputElement>(null);
   const [clearEventsBefore, setClearEventsBefore] = useState<number | null>(null);
   const [searchQuery, setSearchQuery] = useState('');
   const [eventType, setEventType] = useState<CurlEvent['type']>();
-  const allEvents = useRealtimeConnectionEvents({ responseId: response._id });
+  const protocol = response.type === 'WebSocketResponse' ? 'webSocket' : 'curl';
+  const allEvents = useRealtimeConnectionEvents({ responseId: response._id, protocol });
   const handleSelection = (event: CurlEvent | WebSocketEvent) => {
     setSelectedEvent((selected: CurlEvent | WebSocketEvent | null) => selected?._id === event._id ? null : event);
   };
