+++ packages/insomnia/src/ui/hooks/use-realtime-connection-events.ts
@@ -1,8 +1,8 @@
 import { useEffect, useState } from 'react';
 import { useInterval } from 'react-use';
 
-import { CurlEvent } from '../../../main/network/curl';
-import { WebSocketEvent } from '../../../main/network/websocket';
+import { CurlEvent } from '../../main/network/curl';
+import { WebSocketEvent } from '../../main/network/websocket';
 
 export function useRealtimeConnectionEvents({ responseId, protocol }: { responseId: string; protocol: 'curl' | 'webSocket' }) {
   const [events, setEvents] = useState<CurlEvent[] | WebSocketEvent[]>([]);
